Add option to wipe save and validate save CRC before continuing to main menu
If an invalid/corrupt save is found, the user will only be offered the backup/restore options

The wipe save option exists in case the user messes up and restores a save that makes the gameboy crash on bootup. (I don't know if that can actually happen, but just in case)
After all: the backup/restore options allow for restoring saves of different games than the cartridge you're restoring to. (a red save to a blue cartridge, ...)
Therefore the user could also -accidentally- restore a gold save to a blue cartridge (for example).

So yeah, that option is only there to fix theoretic accidents without the user having to take out the battery.