Add "Reset clock" function so you can easily reconfigure your game clock in gen 2
There's no straightforward in gen 2 to reconfigure the in-game clock. There's an arcane key combination that's the worst
on Pokémon Crystal and requires you to calculate some kind of password.

Now PokeMe64 makes it easy: the "Reset Clock" function sets a flag that will let the game prompt you in the main menu to
reconfigure the date/time again.