Initial experiments with a Pokémon Gold Aliexpress reproduction cart from my brother.
This commit adds support for reading the save file of a pokémon gold Aliexpress reproduction cart and copying it
to the carts' SRAM. Note that I specifically said READ. It's the write that will be the hard part.

But it seems like supporting reproduction carts is going to be much harder than I hoped. Let me explain:

What can be easily found online is that these reproduction carts are actually overwriting unused space in the rom to
hold the players' save.

But my current research suggests that I may not be dealing with just 1 or 2 types of reproduction carts.
It may be a whole slew of them, all with different ways to trigger the transfer of SRAM to ROM space, all using
different rom offsets. The latter might even differ between the various editions of the same gen.

Secondly, I had anticipated to check how cartridge flashers such as GBxCart RW are flashing such reproduction carts.
Turns out they use cartridge PIN 31 (AUDIO IN) for that. Problem is: neither the gameboy nor the N64 transfer pak could
programmatically control this pin.

So, these carts have a different way in which the game save mechanism triggers the save transfer. My current
investigation points to these cartridges having a dedicated register to trigger a transfer from an SRAM bank to a ROM
bank. That's my theory based on my disassembly interpretation so far at least. It's untested for the moment.
Problem is: this may be different for every reproduction cart chip type.

For instance, the reproduction cart of Pokémon Gold from my Brother appears to write the value 0x55 to register 0x0555
to trigger the transfer.

It remains to be seen if this could be replicated using the transfer pak. After all, this particular register overlaps
with the RAM/RTC enable register range from the regular MBC3 controller. And the transfer pak can only write in blocks
 of 32 bytes. So we may inadvertedly trigger an MBC3 command when trying to write a single byte to 0x0555.

Anyway, none of the write process is currently handled by these code changes. The changes in this commit just allow
for getting to the gen 2 PokeMe64 main menu with this particular reproduction cart type. It's nowhere near useful yet.

You may be wondering why I'm writing all of this in this commit message. But I just want to write it somewhere. It may
not end up being entirely correct as my investigation continues. But at least this message shows where my mind was at
at this stage.

At this point I'm not even sure if I will succeed in implementing this feature at all. The 32 byte block write hardware
limitation of the N64 transfer pack could potentially become a fatal roadblock.

We'll see. In any case, even if I manage to integrate support for the repro carts I currently have on hand, I'm likely
going to need community support for integrating and testing other types of repro carts.